Force encoding of file coming from open-uri
For some reason the downloaded file is interpreted by open-uri as
being ASCII when it's supposed to be UTF-8 and seems to be correctly
served as that by Google. Forcing the correct encoding makes sure that
encoding flows down to the parsed CSV data and ultimately the SQLite
database.

You'll need to delete your database for this to work, I'm guessing
because the database encoding is set when it's created.